chore(sync): stage all tracked modifications across backend and frontend
Backend domain and service updates, docker-compose, middleware, vite config,
and frontend HTML/JS updates accumulated during amalgamation work.
Includes app/main.py production wiring, defi database layer, websocket hub,
cortex, middleware/security, and all modified frontend public assets.

+++ backend/app/defi/engines/lending.py
@@ -104,7 +104,7 @@ def compound_interest(
     """Return accrued interest (not including principal)."""
     t = elapsed_seconds / 86400 / 365  # fractional years
     n = frequency
-    return principal * ((1 + rate_annual / n) ** (n * t) - 1)
+    return float(principal * ((1 + rate_annual / n) ** (n * t) - 1))
 
 
 def health_factor(collateral_usd: float, outstanding_usd: float, threshold: float = 1.10) -> float:
